Reuse ellipse mesh code for regular polygons
